+++ cluster-autoscaler/cloudprovider/digitalocean/digitalocean_node_group_test.go
@@ -59,15 +59,16 @@ func TestNodeGroup_IncreaseSize(t *testing.T) {
 
 		delta := 2
 
+		newCount := numberOfNodes + delta
 		client.On("UpdateNodePool",
 			ctx,
 			ng.clusterID,
 			ng.id,
 			&godo.KubernetesNodePoolUpdateRequest{
-				Count: numberOfNodes + delta,
+				Count: &newCount,
 			},
 		).Return(
-			&godo.KubernetesNodePool{Count: numberOfNodes + delta},
+			&godo.KubernetesNodePool{Count: newCount},
 			&godo.Response{},
 			nil,
 		).Once()
@@ -162,15 +163,16 @@ func TestNodeGroup_DecreaseTargetSize(t *testing.T) {
 
 		delta := -2
 
+		newCount := numberOfNodes + delta
 		client.On("UpdateNodePool",
 			ctx,
 			ng.clusterID,
 			ng.id,
 			&godo.KubernetesNodePoolUpdateRequest{
-				Count: numberOfNodes + delta,
+				Count: &newCount,
 			},
 		).Return(
-			&godo.KubernetesNodePool{Count: numberOfNodes + delta},
+			&godo.KubernetesNodePool{Count: newCount},
 			&godo.Response{},
 			nil,
 		).Once()